pulled get methods into its own repo
